% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme_S3.R
\name{print.lpmec_moderator}
\alias{print.lpmec_moderator}
\title{Print method for lpmec_moderator objects}
\usage{
\method{print}{lpmec_moderator}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lpmec_moderator}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Print method for lpmec_moderator objects
}
