% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme_S3.R
\name{print.lpmec_moderator_onerun}
\alias{print.lpmec_moderator_onerun}
\title{Print method for lpmec_moderator_onerun objects}
\usage{
\method{print}{lpmec_moderator_onerun}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lpmec_moderator_onerun}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Print method for lpmec_moderator_onerun objects
}
