% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme_S3.R
\name{print.lpmec_multivariate}
\alias{print.lpmec_multivariate}
\title{Print method for lpmec_multivariate objects}
\usage{
\method{print}{lpmec_multivariate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lpmec_multivariate}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Print method for lpmec_multivariate objects
}
