% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme_S3.R
\name{print.lpmec_reliability_bounds}
\alias{print.lpmec_reliability_bounds}
\title{Print method for lpmec_reliability_bounds objects}
\usage{
\method{print}{lpmec_reliability_bounds}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lpmec_reliability_bounds}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object \code{x}, returned invisibly.
}
\description{
Print method for lpmec_reliability_bounds objects
}
