% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme_S3.R
\name{summary.lpmec_multivariate}
\alias{summary.lpmec_multivariate}
\title{Summary method for lpmec_multivariate objects}
\usage{
\method{summary}{lpmec_multivariate}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lpmec_multivariate}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A data frame of aggregated latent-predictor coefficient estimates.
}
\description{
Summary method for lpmec_multivariate objects
}
