% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpme_S3.R
\name{summary.lpmec_multivariate_onerun}
\alias{summary.lpmec_multivariate_onerun}
\title{Summary method for lpmec_multivariate_onerun objects}
\usage{
\method{summary}{lpmec_multivariate_onerun}(object, ...)
}
\arguments{
\item{object}{An object of class \code{lpmec_multivariate_onerun}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A data frame of latent-predictor coefficient estimates.
}
\description{
Summary method for lpmec_multivariate_onerun objects
}
